chore: split oversized files under 150-line limit
- RouteMessageUseCase (220→116): extract ReceptionistRouter (131)
- AgentLoopHelpers (209→93): extract AgentLoopTypes (53), AgentLoopBuilders (72)
- LifecycleUseCase (172→92): extract StatsGenerator (98)

All 381 tests pass.
