+++ src/application/query/AgentLoopHelpers.ts
@@ -1,9 +1,6 @@
 import type { AIMessage, AIContentBlock, ProviderPlugin } from '@supaproxy/providers'
-import type { ExecutionRailRegistry, RetrievalRailRegistry } from '@supaproxy/guardrails'
-import type { AuditLogData } from '../../domain/audit/repository.js'
-import type { ManageConversationUseCase } from '../conversation/ManageConversationUseCase.js'
-import type { ToolEntry, ToolCallRecord } from './ToolCallProcessor.js'
 import type { ToolCallProcessor } from './ToolCallProcessor.js'
+import type { AgentLoopConfig, AgentLoopResult } from './AgentLoopTypes.js'
 import { IS_PRODUCTION } from '../../config.js'
 import { DEFAULT_MAX_RESPONSE_TOKENS } from '../../defaults.js'
 import pino from 'pino'
@@ -12,57 +9,6 @@ const log = pino({ name: 'agent-loop' })
 
 const NO_RESPONSE_MESSAGE = '(no response)'
 const MAX_ROUNDS_MESSAGE = 'Ran out of tool-call rounds. Please simplify your question.'
-
-export interface AgentLoopConfig {
-  model: string
-  systemPrompt: string
-  maxToolRounds: number
-  tools: ToolEntry[]
-  history: Array<{ role: 'user' | 'assistant'; content: string }>
-  apiKey: string
-  workspaceId: string
-  conversationId: string
-  executionRails: ExecutionRailRegistry | null
-  retrievalRails: RetrievalRailRegistry | null
-}
-
-export interface AgentLoopResult {
-  answer: string
-  toolsCalled: ToolCallRecord[]
-  connectionsHit: string[]
-  tokensInput: number
-  tokensOutput: number
-  costUsd: number
-  durationMs: number
-  error: string | null
-}
-
-export interface QueryMeta {
-  consumerType: string
-  channel?: string
-  userId?: string
-  userName?: string
-  conversationId?: string
-  sessionId?: string
-  systemPromptOverride?: string
-  skipTools?: boolean
-  routedFrom?: string
-  routedFromConversationId?: string
-  priorHistory?: Array<{ role: 'user' | 'assistant'; content: string }>
-}
-
-export interface QueryResult {
-  answer: string
-  toolsCalled: string[]
-  connectionsHit: string[]
-  tokensInput: number
-  tokensOutput: number
-  costUsd: number
-  durationMs: number
-  error: string | null
-  conversationId: string
-  sessionId: string
-}
 
 export async function runAgentLoop(
   query: string,
@@ -142,67 +88,6 @@ export async function runAgentLoop(
   return result
 }
 
-export function buildEmptyResult(overrides: Partial<AgentLoopResult> = {}): AgentLoopResult {
-  return { answer: '', toolsCalled: [], connectionsHit: [], tokensInput: 0, tokensOutput: 0, costUsd: 0, durationMs: 0, error: null, ...overrides }
-}
-
-export function buildQueryResult(partial: Partial<QueryResult> & { answer: string; conversationId: string; sessionId: string }): QueryResult {
-  return {
-    toolsCalled: [],
-    connectionsHit: [],
-    tokensInput: 0,
-    tokensOutput: 0,
-    costUsd: 0,
-    durationMs: 0,
-    error: null,
-    ...partial,
-  }
-}
-
-export function buildAuditLogData(
-  auditLogId: string,
-  workspaceId: string,
-  conversationId: string,
-  query: string,
-  result: { toolsCalled: ToolCallRecord[]; connectionsHit: string[]; tokensInput: number; tokensOutput: number; costUsd: number; durationMs: number; error: string | null },
-  meta: QueryMeta,
-  screening?: { screeningAction: string | null; screeningCategories: string[] | null; screeningMs: number | null },
-  knowledgeChunks?: number,
-): AuditLogData {
-  return {
-    id: auditLogId,
-    workspace_id: workspaceId,
-    conversation_id: conversationId,
-    consumer_type: meta.consumerType,
-    channel: meta.channel || null,
-    user_id: meta.userId || null,
-    user_name: meta.userName || null,
-    query,
-    tools_called: JSON.stringify(result.toolsCalled.map(t => t.name)),
-    connections_hit: JSON.stringify(result.connectionsHit),
-    tokens_input: result.tokensInput,
-    tokens_output: result.tokensOutput,
-    cost_usd: result.costUsd,
-    duration_ms: result.durationMs,
-    error: result.error,
-    input_screening_action: screening?.screeningAction || null,
-    input_screening_categories: screening?.screeningCategories ? JSON.stringify(screening.screeningCategories) : null,
-    input_screening_ms: screening?.screeningMs || null,
-    knowledge_chunks_used: knowledgeChunks || 0,
-  }
-}
-
-export async function recordMessages(
-  conversationUseCase: ManageConversationUseCase,
-  conversationId: string,
-  query: string,
-  answer: string,
-  auditLogId: string,
-): Promise<void> {
-  try {
-    await conversationUseCase.recordMessage(conversationId, 'user', query)
-    await conversationUseCase.recordMessage(conversationId, 'assistant', answer, auditLogId)
-  } catch (err) {
-    log.error({ error: (err as Error).message }, 'Failed to record conversation messages')
-  }
-}
+// Re-export types and builders for backwards compatibility
+export type { AgentLoopConfig, AgentLoopResult, QueryMeta, QueryResult } from './AgentLoopTypes.js'
+export { buildEmptyResult, buildQueryResult, buildAuditLogData, recordMessages } from './AgentLoopBuilders.js'
